Add `DbRand` to allow injectable randomness (#611)
This PR is a follow-on to #610. I've added a `DbRand` struct to
`rand.rs`. The intent is to provide the equivalent to
`rand::thread_ng()`, but in an object that can be put in an Arc to share
everywhere, and which doesn't use locking. The code achieves this using
the `thread_local` crate; we create a new random number generator (RNG)
once for each thread. To seed the thread RNGs, we use an AtomicU64,
which we increment after each use.

The `RngAlg` uses `Xoroshiro128PlusPlus`, which is portable (runs
deterministically across environments and OS's). `Xoroshiro128PlusPlus`
also uses splitmix to safely randomize the u64 and maintain entropy.

This PR does _not_ update the UUID/ULID generation to use `DbRand`. I
plan to do that in the next PR. I'm trying to keep things readable.