Add `GarbageCollectorBuilder` and make `GarbageCollector` non-static (#613)
This is the next follow-on after #610. In this PR, I introduce a
`GarbageCollectorBuilder` and make the `GarbageCollector` non-static. I
feel it's much cleaner. It also lays the groundwork to be able to use
`DbRand` to generate `Uuid`s and `Ulid`s. When I started doing that
work, I found that there were a ton of invasive changes in
admin/compaction/gc because of all these static methods passing params
around.