feat(wal): add a wal buffer manager; support dirty read semantics in RFC-008 (#519)
## motivation

In the current design, the WALs function as mini-memtables. When a WAL
table is full, it is frozen into an immutable WAL table, flushed to the
object store, and the writes are then applied to the MemTable.

For users who want to read writes immediately, we allow reading from the
data WAL to enable the dirty read semantic. Users can also specify
`durability_filter: DurableLevel::Remote` to ensure that only durable
data is read.

In RFC-008, we proposed to support the sync commit semantic, which is a
precondition of supporting transactions.

## what this pr changes

This PR is not considered as a full implementation of RFC-008 but makes
RFC-008 trivial to be added.

This PR modifies the interaction between the WAL and MemTable as
follows:

1. All writes will be written to the WAL first (which is literally what
Write Ahead means).
2. and applied to the MemTable immediately (instead of waiting for the
WAL to be full).

After this change, readers will only access the MemTable, and the WAL
will not be accessed unless replaying during startup.

For the dirty read semantic, users can read data from the MemTable
without the limitation of `last_committed_seq`.

For the durable filter, users can read data up to the
`last_flushed_seq`.

Consequently, this PR makes the following changes in the codebase:

1. Removed WAL-related logic from the read path.
2. Apply writes to the MemTable immediately after appending to the WAL.
3. Track the `last_committed_seq` and `last_persisted_seq`.
4. Extracted a WAL buffer manager to encapsulate the details about
in-memory WAL management and flushing.

## WalBufferManager

`WalBufferManager` is an experiment to encapsulate the details about
in-memory WAL management and flushing, while reducing the burden of
state management from the `DB` struct.

It hopes to expose a minimal set of methods:

* `append()` and `maybe_trigger_flush()`: used for appending to the WAL
during commit.
* `flush()`: Flushes all unflushed WALs and awaits their persistence
when `sync` is enabled.

It also provides a retry mechanism for the flushing operation on
temporary I/O errors.

Also, when we add local storage flushing in the future, all the changes
can be placed within the `WalBufferManager` without affecting any other
parts of the codebase.

## future works

- enable `sync` option in the write side
- expose metrics around WAL flushing
- more test cases around `WalBufferManager`

---------

Co-authored-by: Chris <criccomini@users.noreply.github.com>