Add `CompactorBuilder` and make `Compactor` async (#620)
This is the next DST PR after #613. The compactor was originally a
mixture of sync and async code. The event loop was synchronous and ran
in a thread. This differs from the GC, which has an async event loop.
This PR unifies the design of the compactor and the garbage collector.
My hope is that this will make it easier for us to do DST since we can
have the compactor run on the same single threaded tokio runtime as
everything else.

- Replace crossbeam with tokio
- Remove `CompactorOrchestator` (it seemed like an unneeded layer of
abstraction)
- Replace the thread in `Compactor::new` with `start_async_task` and
`start_in_bg_thread`
- Add `CompactorBuilder` to `builder.rs`
- Get rid of err_*x and replace `CompactorMainMsg` with a
`CancellationToken`
- Update a bunch of tests

Notes:

1. I opted to leave the `CompactionExecutor` trait synchronous. The
`TokioCompactionExecutor` remains as it was before: it receives a tokio
runtime handle, which it uses for async code. I'm open to making it
async, but not in this PR. Would be interested in everyone's thoughts.
2. I found some flaky tests. The first were a group in
`compactor_state.rs`; these were fixed in #619. The second was
`test_should_read_from_compacted_db` in `db.rs`. This one also appeared
to be depending on compactor timing. I added a failpoint to pause the
compactor loop until all data is written and everything works. I did
some debugging and it seems like the compactor was running fast enough
now that it was compacting stuff in such a way that it would get some
L0's compacted and into an SR and then reach a steady state. **Please
pay special attention to this test. I want to make sure I didn't break
any assumptions in my refactoring.**