fix(wal): track db fatal when wal flusher got unrecoverable error (#631)
this is another try of having fatal state to get propogated when wal
flushing got unrecoverable error.

this `handle_fatal` is kinda ugly, i suppose another way is directly
expose `db.state.error` to underlying components like
`WalBufferManager`, as it can be considered as a globally coordinated
primitive, instead of an inner state of db state.

---------

Co-authored-by: Chris <criccomini@users.noreply.github.com>