Use `DbRand` to generate UUIDs and ULIDs (#633)
This is the next PR after #620 in the enable-DST series. Here, we
replace static UUID and ULID generation with implementations that use
`DbRand`. This will make UUID and ULID generation deterministic if we
seed the DB.

This work depended on the compactor refactoring in #620 because the
compactor used ULIDs and UUIDs a lot. Prior to #620, it had no builder
and was mostly static. Updating each static method bled the parameter
updates everywhere. Now that the compactor has a `self` it can reference
for most methods, adding `DbRand` is much simpler.

We now have `gen_uuid` and `gen_ulid` methods that are automatically
added to any `RngCore` when `utils::IdGenerator` is imported. Both
methods use the `RngCore` it's attached to for ULID/UUID generation.

Most of the changes are fairly mechanical. A couple of things to watch
out for:

- I replaced `FenceableManifest ::write_checkpoint`'s `Option<Uuid>`
with a required `Uuid`. This moved random number generation out of the
`write_checkpoint` method, which seemed much cleaner than passing in a
`DbRand`.
- `CompactorState::submit_compaction` now requires that a `Uuid` is
passed in. Previously, it was doing UUID creation itself. Again, this
seemed cleaner.
- It seems to me that the `Compaction` ID is also used as the
`CompactionJob` id in `CompactorEventHandler::submit_compaction`. I
maintained that logic, but it would be nice to confirm this what's
supposed to happen.