test: let test_get_after_put to use explicit options instead of defaults (#652)
related: #647

this pr uses explicit options on write options and read options, to test
over the case about a `get()` should always get the committed data with
`dirty: false`.

(currently the `sync` field is not supported yet. all the writes' sync
level are considered as memory, this should be added in a pr).