Default to Foyer cache and 256MiB `max_sst_size` (#655)
@Barre discovered that Moka could get into a ton of contention under his
workload. It's not immediately apparent what the issue is with Moka, but
switching to Foyer fixed the problem immediately.

@Barre also discovered that the 1GiB default for compacted SSTs can lead
to a lot of write amplification and slow things down. 256MiB was a much
better default for him. In general, I think that makes sense for others,
too. The cost is 4x the number of SSTs listed in the manifest and on
list() calls to `compacted`. If that's an issue for users, they can set
their SST size to 1GiB.

Fixes #602, #646