feat(disk_cache): implement extensible cache admission framework (#648)
This is the first step towards a solution for #198, which implements a
extensible cache admission framework. We introduced an `AdmissionPicker`
to manage all admission policies.

Users can customize `AdmissionPolicy` or directly use the admission
policies we provide. Currently we have implemented an admission policy
that accepts all cache entries, so this will not affect the current disk
cache write process.