Add compactor progress (#666)
I added a `CompactionProgressTracker` to `compactor.rs`. I updated the
compactor to log the compaction progress along with the compaction state
when the ticker fires in the event loop.

Here's what it looks like:

```
2025-07-02T22:42:44.903438Z DEBUG slatedb::compactor: compaction progress id=0a1e6fbe-c117-4c52-b9cf-6f9af317bb96 current_percentage="2%" processed_bytes=13802880 estimated_total_bytes=540058533
```

This will give wildly inaccurate results when compression is used since
we're using compressed bytes when estimating size, but uncompressed
bytes when getting the size of key/value pairs. IMO, it's still useful
for tracking raw byte progress. Once we store actual byte or row count
information in metadata, we can use it here to calculate accurate
progress.