fix: relax the bound of the flush channel in wal_buffer; add logs on flushing failure (#660)
currently the WAL buffer uses a channel with a bound of 1. this may lead
to issues when the flush task blocks for an extended period.

in write_batch.rs, `maybe_trigger_flush().await?` is called on every
write. `maybe_trigger_flush` sends a message to the flush channel if it
determines a flush should be triggered.

however, while the background flushing task is busy saving SST files to
the object store, no one receives messages from the channel, which then
blocks the sender. if users have many tokio tasks performing write
operations simultaneously, this could block a lot number of user tasks.

using an unbounded channel would make `maybe_trigger_flush().await` more
asynchronous and less likely to block user writes.

(it would be good to monitor the length of this unbounded channel, but i
think it can be addressed in a separate PR)

---------

Co-authored-by: Chris <criccomini@users.noreply.github.com>