Protected against key/value length overflows in `SstRowCodecV0` (#679)
SlateDB's block format only allows u16::MAX sized keys[^1] and u32::MAX
sized values.

The `Db` currently allows users to put() with keys that are > u16 and
values that are > u32. The `SstRowCodecV0` happily accepts these rows
and casts the lengths (which are `usize` for developer ergonomics) with
`as u16` and `as u32`. This will silently overflow lengths. The outcome
is that future decodes will fail in odd ways (I was seeing `Error:
InvalidRowFlags { encoded_bits: 244, known_bits: 15, message: "Unable to
parse flags. This may be caused by reading data encoded with a newer
codec." }`).

This PR adds protection both at the `WriteBatch` level and the
`SstRowCodecV0` level. It also adds tests to validate things work as
expected. I opted to do the validation as a `panic` (via `assert`) in
the `WriteBatch` `put`/`delete` calls since a) it's in keeping with our
current `key.is_empty()` check, and b) we don't expose return a `Result`
for the methods (I didn't want to do a bw-incompatible change for this).

I also briefly considered doing the check in `batch_write.rs`, but that
would a) burn CPU on the writer thread, b) defer the failure until
`write` is called rather than exposing the issue immediately on
`batch.put()`, and c) could leave us with a partially written batch
unless we fully validate before writing (c.f. cpu burn). The advantage
would be that we could expose a error as a `SlateDbError` rather than
panicking. I don't the tradeoffs are worth it.

[^1]: _Technically_ we allow up to a max of u16 prefix bytes and u16
suffix bytes. But since the prefix can always be 0, we have to cap the
sum of the two at 16.