Record bytes_compacted correctly. (#682)
When a writer is finished, the compaction process is recording 0 as the
size, instead of the size of the current bytes processed until that
point.

I was looking at these two lines while trying to implement #676, and it
seems wrong. Since this is a small change, I rather get it out while I
try to figure out how to implement #676 correctly.