`await_durable` writes hang when `wal_enabled` is false (#680)
A user that disables the WAL and then calls a single write() with
`await_durable` will block forever. This is because we only freeze (and
eventually flush) the active memtable when `l0_sst_size_bytes` is
exceeded. Unless the first `write()` call is > `l0_sst_size_bytes`, the
write will hang. I don't think we've come across this yet because most
people write() with `await_durable` set to false, then flush later (or
they write so much they exceed the memtable byte threshold frequently
enought not to notice long pauses).

After some discussion, I've opted to leave the behavior as-is. A small
change is introduced to `batch_write.rs` in this PR. The change adds a
monitor to the first write if the WAL is disabled. If the write takes
longer than 5 seconds, this warning is printed:

```
warn!("First write not durable after 5 seconds and WAL is disabled. \
SlateDB does not automatically flush memtables until `l0_sst_size_bytes` \
is reached. If writer is single threaded or has low throughput, the \
applications must call `flush` to ensure durability in a timely manner.");
```