feat(compaction): add retention policy for snapshot versioned data (#638)
lsm engines like leveldb/rocksdb mostly takes a reference counting
approach to manage the active snapshots.

in slatedb, we have a single writer / multiple reader architecture, as a
result, we can not track the active snapshots in a single process. to
allow having Snapshots or TimeTravel, we can consider having a
time-retention based approach, like:
https://www.tabular.io/apache-iceberg-cookbook/data-operations-snapshot-expiration/

this pr plans to add a PoC about this time-based retention support.

it should make some modifications over the compaction phase. currently,
the compaction phase is:

1. we use MergeIterator to iterate the latest versioned key/values.
2. filter out the expired/tombstoned key/values (if it's in last level).
3. append to the new ssts.

with retention in consideration, it will become:

1. iterate the multiple versions of a single key in the decreasing order
of seqnum.
2. if a key has the multiple versions, filter out the versions whose
create_time is earlier than the retention time.
3. if a key has only one version, and expired/tombstoned, skip this
key/value (if it's last level)
4. append to the new ssts.

to manage this retention logic, this pr adds a RetentionBuffer for this,
it works as a state machine to:

1. collect the multiple versions of the same key
2. when all the versions has been collected, handle the retention policy
3. after retention policy as applied to the versions, pop one version at
a time

this algorithm is not optimal on memory, but it's easier to reason about
the correctness of implementation. (we can do optimizations after we
reviewed that this handling logic has no issues.)

after this change, the expiry & tombstone handling in compactor.rs also
has to be multi-version aware, so this PR moves the expiry & tombstone
handling logic into retention_iterator.rs as well.