Remove tokio::time usage (#683)
I thought it would be OK to continue to use `tokio::time` directly in
SlateDB since we can pause/resume it in tests (e.g. with
`start_paused=true`. Unfortunately, I found an issue with that when
working on some deterministic simulation tests. The issue arises if you
want to run a simulation test twice as though it were the same point in
time. Unfortunately, Tokio's `time` module does not allow users to set a
time in the past, only to advance it.

This PR goes through the code to replace (almost) all usage of
`tokio::time` methods and types with our `SystemClock` trait. To
accomplish all that tokio does for us, I had to introduce some new
methods to our system clock: advance, sleep, and ticker. I opted to
replace`timeout` usage witha `tokio::select!` rather than fight with
Rust's lifetime stuff more. If people strongly object lacking a
`timeout` method, I can revisit it.

There are a few places where `tokio::time` is still used. All are in
tests, and they're fairly harmless. Some day maybe we can scrub the
remaining ones, but this PR is already quite large.

This PR also introduces `MockSystemClock`, which is a mock clock that's
completely divorced from Tokio's time as well as std::time. It's
entirely self-encapsulated, and allows users to set any time they want.
This is the clock I use in my DST branch, and it seems to work quite
nicely.

I've also added tests for both the default and mock system clocks.

Note that the mock clock is only available when SlateDB's `test-util`
feature is enabled (this is a new feature in this PR). I had to go this
route because DST is done in the `slatedb-dst` crate, which only has
access to public APIs in SlateDB. Otherwise, I would have simply used
`#[cfg(test)]`. The `test-util` follows Tokio's pattern as well.