Implement the Public Error API. (#697)
This is an implementation of RFC 7, but with caveats.

When I started implementing it, I realized that it lacked some details
on the design itself. I opted for changing the `Error` type definition
to match what I've seen in other very popular crates, instead of using
an enum to define all the errors, the `Error` type is a generic error
that includes a `ErrorKind` enum with the different error variants. This
is similar to the implementation that the AWS SDK follows, for example.

I also added a section to the RFC about internal vs external errors, and
how to format errors to be consistent across the board. We can use that
as a guideline when we introduce new errors, I noticed that there were a
lot of variations in format, and I think it's better to constraint
ourselves to some clear guidelines.

This is a short summary of what this massive PR includes:

- Update the RFC with more information about error types and formatting.
- Create the new public Error type.
- Create a transformation between SlateDBError and the new Error type.
- Update all public endpoint to use the new Error type.
- Update abstract internal errors to be more specific.

Fixes #508