Replace `tracing::` logs with `log::` logs in `slatedb` (#702)
We should be using the `logs` facade crate everywhere in `slatedb` since
it's a library. I added clippy macro checks to enforce this and updated
the code accordingly.

I left the tracing_subscriber (vs. env_logger or others) for output in
tests and tools since that's what `opendal` does and I didn't want to
overthink things. I also left the -cli and -bencher crates to continue
using `tracing` for their logs since they're applications, not
libraries.

I also took the opportunity to prevent...

- stdout print's
- Option::unwrap
- Result::unwrap