Add deterministic simulation tests to SlateDB (#694)
A `slatedb-dst` package is added in this PR. The package contains
dst.rs, which contains a deterministic simulator that runs random
operations against a SlateDB instance and verifies that the DB has the
correct data. Several integration tests are available in
`simulation.rs`; they've been added to the `pr.yaml` check, too. See the
README.md, dst.rs docs, and simulations.rs docs for usage and
implementation details.

There is still much to be done to improve the simulation tests:

- Add unit tests to dst.rs to verify method behavior.
- Add a CLI to trigger DST's outside of `cargo`
- Incorporate DSTs into the `nightly.yaml`
- Complete the `TODO`s to add more randomness and configuration
- Add TTL, checkpointing, DbReader, and fencing (multi-writer), and DB
close/open checks and actions
- Incorporate the object store file cache and inject faults in the
filesystem
- Inject faults in the object store requests
- Evaluate using proptest to build DB settings and generate random data
- Verify log output is identical in `test_dst_is_deterministic` (we only
verify final seed and times are identical)

Still, I feel this is a good start and the PR is already fairly lengthy.
I'd like to do the CLI and nightly.yaml as follow-ons. The remaining
items can be completed over time.

While running the tests, I discovered two instances of non-determinism
in SlateDB:

- The compactor was using `spawn_blocking` to start and stop compactions
with the executor. I assumed `spawn_blocking` would be deterministic,
but it turns out not to be. It spawns outside the `current_thread` Tokio
runtime and creates a non-deterministic race between `.await` calls in
the `current_thread` and those in the spawned thread. I fixed this by
adding a branch for `#[cfg(dst)]` that uses tokio's `spawn`, which runs
on the runtime threads. I opted not to use `spawn` for production code
because we can't trust pluggable executors not to block for a long time
on start/stop compaction calls; they could block our runtime threads.
- ULID was using the RNG to deterministically generate the random
portion of its bytes, but was still using its own clock for the
timestamp portion. I fixed this by adding a `timestamp_ms` field to
`gen_ulid`.

I found the following resources helpful when doing this work:

- https://pierrezemb.fr/posts/tokio-hidden-gems/
- https://www.reddit.com/r/rust/comments/1jr8ogo/comment/mle11wh/
- https://s2.dev/blog/dst

Fixes #267