feat(snapshot): add retention min seq to manifest (#701)
as discussed in #638, we plan to track the `retention min seq` in the
manifest during running transactions, and let the compactor get the
`retention min seq` via the manifest, so the compactor can safely
cleanup the data not actively accessed by snapshots/transactions.

this PR adds a `retention_min_seq` field to the manifest, and enables
this field for compactor. however, this pr is expected not to change the
existing behavior, until we add logics to flush this field to manifest.

this PR can also be considered as a subtask extracted from #688. #688
should later add a method to flush the `retention_min_seq` to the
manifest in the background.

---------

Co-authored-by: Chris <criccomini@users.noreply.github.com>