Add a nightly DST test (#733)
This PR integrates the DST into the nightly Github workflow.

- Adds a `test_dst_nightly` test to simulation.rs.
- Adds a `slow` cfg to `Cargo.toml` to signal slow tests should be run.
- Updates `nightly.yaml` to run the DST on a WarpBuild instance.
- Adds a `DstDuration` param that can be either `Iterations` or
`Duration` (wall-clock time) to the `run_simulation`. This allows us to
run simulations for a set amount of time rather than a set number of
iterations.
- Update `nextest.toml` to include a new `dst-nightly` profile.
- Update docs and README.md

`test_dst_nightly` runs a separate DST on each core of the host machine.
The tests run in parallel for 50 minutes.

Future improvements:

- Fail as soon as a handle finishes rather than joining them
sequentially.
- Make `test_dst_nightly` use an environment variable so the 50m
duration isn't hard-coded.