Use LocalFileSystem for nightly DST (#738)
This PR adds a new environment variable for DSTs called
`SLATEDB_DST_ROOT`. The `slow` DST uses this as the root to create a
`LocalFileSystem` `object_store` for each `Db`. Previously, we were
using `InMemory`, which was accruing a ton of memory because GC wasn't
running very frequently and we were creating a ton of garbage in the
test.