Fix documentation issue for garbage collection: "poll_interval" -> "interval" (#742)
It looks like the garbage collector options were changed from
"poll_interval" to just "interval", but docs were not updated to
reflect.

The compactor still uses "poll_interval", so care was taken to _not_
update the docs. Searched through code for instances of `poll_interval`,
determined if it was for the garbage collector or compactor, and
replaced as appropriate.