Limit nightly DST to 12 minutes (#753)
We are running out of disk space in our nightly DSTs. The reason appears
to be that the `LocalFilesystem` `object_store` uses the actual
timestamps from the machine time. Our DST mock system clock starts at
the UTC epoch, though. Consequently, all the files have a last modified
time _way_ ahead of our mock system clock. This causes the GC not to
ever remove anything.

I plan to open a GH issue to figure out what to do about this. In the
meantime, I'm limiting the nightly DST execution time to 12 minutes so
we don't run out of disk on the WarpBuild machine.