[2/4] implementation of gorilla-based single tier sequence tracker (#737)
See #710 for the full context and how this is wired in.

This PR adds a data structure that can help track a map of sequence
number -> timestamp with lossy granularity.

Some notes:

1. It encodes using the Gorilla encoding strategy. Unfortunately I
couldn't find a well maintained rust implementation of it, and it isn't
too much code to just vendor it so I implemented it here.
2. Gorilla recommends storing timestamps with second granularity to
reduce the delta-of-deltas. I do that here by converting all timestamps
to seconds. This is probably fine as this map is intended to be used for
operational use cases.
3. The current implementation only supports a single tier with even
spacing. When it fills up it downsamples everything. This is likely not
the ideal final behavior as we likely want newer datapoints to be stored
with higher granularity, which is what I'm intending to do with the
tiered approach.