feat: add DbSnapshot (#688)
this pr aims to add the API for creating snapshot.

a `DbSnapshot` contains all the read-only operations in a `Db` object.
it records the seqnum at the moment when the `DbSnapshot` is created,
and always making a consistent view as long as this `DbSnapshot` object
lives.

in the current design, `DbSnapshot` is not expected to work on the db
read instance.

we can divide this pr into the following subtasks:

- [x] scan() get() with `seq` passed
- [x] track all the living `DbSnapshot` objects (maybe in a
`TransactionManager`?)
- [x] extract the `min_retention_seq` from living `DbSnapshot` objects,
and persist it in manifest (flushed whenever memtable flushes)
- [x] let the compactor handle the `min_retention_seq` by
RetentionIterator #701
- [x] exposing the API
- [x] add tests