Apply batch operations in DST after writing to the database. (#775)
Currently, the state of the DST is modified before a write batch hits
the database. This could create inconsistencies where the state changes,
but the database doesn't change because there was an error writing to
it.

This change the order of operations to apply the write batch to the
state only after the write to the database successfully completes.