Shutdown cleanly in `test_concurrent_writers_and_readers` (#807)
It looks like we were not doing a clean shutdown in
`test_concurrent_writers_and_readers`. Instead, we were calling
`abort()` on the reader threads and then ignoring errors returned by the
reader tasks.

This PR updates the code to use a cancellation token to shutdown cleanly
and then verify the readers have no errors.

Fixes #806