feat(txn): Add conflict check for SI transaction (#805)
fixes #787

this pr prepared the conflict checking logics in the implementation of
SI transaction.

this pr should not affect all the exisiting behaviour: when there's no
active non-readonly transactions, no `recent_committed_txn` is tracked
in memory at all.

active non-readonly transaction is only possible to exists after
`DbTransaction` interface is introduced, which is planned in a seperated
issue #816.

- [x] prepare the internal api in txn manager.
- [x] extensive unit tests inside txn_manager itself.
- [x] proptest in txn_manager.
- [x] integrate it inside commit pipeline
- [ ] ~expose a `DbTransaction` interface~

---------

Co-authored-by: Chris <criccomini@users.noreply.github.com>