Compaction persistence rfc (#695)
This PR provides a RFC for the feature of persisting compaction state to
the object store. It discusses separating out a `CompactionState` from
`Manifest`. It also discusses mechanisms of leveraging the persistence
of compaction state to enable retry attempts to failed compactions,
better tracking of `CompactionJob` and cost implications of persisting
the compaction state.

See #673 for more details.